Solucionats errors amb els models, controllador, i afegit scope a Program

+++ DNDAPI/Models/SubModels/ExpertiseOptionFeature.cs
@@ -6,7 +6,7 @@ namespace DNDAPI.Models.SubModels
     public class ExpertiseOptionFeature
     {
         [BsonElement("choose")]
-        public int? Choose { get; set; }
+        public double? Choose { get; set; }
 
         [BsonElement("from")]
         public ExpertiseOptionFromFeature[]? From { get; set; }
